Splatoony UI Update (#14)
* Revamp buttons to mimick Splatoon 2's UI.

* Adds hover/focus animations to buttons.
* Moves primary button to the right.
* Adds a missing closing </div> tag.
* Reorder CSS so Bootstrap is loaded first, then our styles.
* Create a dedicated CSS file for buttons/dialog/modals.
* Blurs background when a modal is open.

* Add "leak" animation when clicking dialog buttons.

* Prevent footer jumping around on modal dialog.

It is a result of the missing </div> tag, the position was being
determined by the body tag.

As a trade off, the footer will be no longer "sticky".
(Personally I'm OK for this type of footer)

* Delay closing of AngularJS modal dialogues.

Gives enough time for the button "clicked" animation to play.

* Style the scrollbar (for WebKit browsers)

* Revamp presentation of dialog boxes.

* Adds the "hanger" to the top.
* Aligns modal in the center.
* Update animation to mimick Splatoon 2's UI.

* Minor alignments to margins/heights.

For main page, due to changes made to card header.

* Add custom styling for dropdown menu.

* Add manifest and mobile icons.

This is to improve integration on platforms that support it. E.g. Chrome
for Android.